A list object of class 'mbal_gas' for material balance analysis

Create an object of class 'mbal_gas'

mbal_param_gas(
  input_unit = "Field",
  output_unit = "Field",
  fluid = "gas",
  aquifer_model = NULL,
  G = NULL,
  phi = NULL,
  swi = NULL,
  Gp = NULL,
  Np = NULL,
  Wp = NULL,
  Wi = NULL,
  We = NULL,
  pd = NULL,
  p = NULL,
  pvt = NULL,
  cf = NULL,
  M = NULL,
  phi_a = NULL,
  perm_h_a = NULL,
  perm_v_a = NULL,
  h_a = NULL,
  r_a = NULL,
  r_R = NULL,
  w_a = NULL,
  l_a = NULL,
  tetha = NULL,
  muw_a = NULL,
  cw_a = NULL,
  cf_a = NULL,
  wf = NULL,
  sgrw = NULL
)

Arguments

input_unit
a unit system for parameters, only the character string 'Field' is accepted

output_unit
a unit system for properties, only the character string 'Field' is accepted

aquifer_model
defaulted to NULL, otherwise must be a character string, one of the following eight options: 'uss_rad_edge', 'uss_rad_bottom', 'uss_lin_edge', 'uss_lin_bottom', 'pss_rad_edge', 'pss_lin_edge', 'pss_lin_bottom', 'pot'. For further information about each model, please see 'Raquifer' package reference manual (https://cran.r-project.org/web/packages/Raquifer/index.html)

G
original gas in place, SCF.

phi
reservoir porosity, a numeric fraction

swi
initial water saturation in reservoir, a numeric fraction

Gp
cumulative gas production, SCF

Np
cumulative oil production, STB

Wp
cumulative water production, STB

Wi
cumulative water injection, STB

We
cumulative aquifer water influx, BBL. If unknown, a NULL value must be assigned

pd
dew point pressure, a numeric value, psi

p
reservoir pressure, a numeric vector, psi

pvt
a data frame of PVT properties including pressure 'p' in 'psi', oil formation volume factor 'Bo' in 'bbl/stb', solution gas-oil ratio 'Rs' in 'scf/stb', oil viscosity 'muo' in 'cp', volatilized oil-gas ratio 'Rv' in 'stb/scf', gas formation volume factor 'Bg' in 'bbl/scf', gas viscosity 'mug' in 'cp', water formation volume factor 'Bw' in 'bbl/stb', and water viscosity 'muw' in 'cp'

cf
formation compressibility, a numeric value or vector, 1/psi

M
ratio of non-net-pay pore volume to the reservoir (net-pay) volume, a numeric fraction.

phi_a
aquifer porosity, a numeric fraction

perm_h_a
aquifer horizontal permeability, md. Used in 'uss_rad_edge', 'uss_rad_bottom', 'uss_lin_edge', 'pss_rad_edge', 'pss_lin_edge' and 'pot' aquifer models

perm_v_a
vertical permeability, md. Used in 'uss_rad_bottom', 'uss_lin_bottom', 'pss_rad_bottom', and 'pss_lin_bottom' aquifer models

h_a
aquifer height, ft

r_a
aquifer radius, ft. Used in 'uss_rad_edge', 'uss_rad_bottom', 'pss_rad_edge', and 'pot' aquifer models

r_R
reservoir radius, ft. Used in 'uss_rad_edge', 'uss_rad_bottom', 'pss_rad_edge', and 'pot' aquifer models

w_a
aquifer width, ft. Used in 'uss_lin_edge', 'uss_lin_bottom', 'pss_lin_edge', and 'pss_lin_bottom' aquifer models

l_a
aquifer length, ft. Used in 'uss_lin_edge', 'uss_lin_bottom', 'pss_lin_edge', and 'pss_lin_bottom' aquifer models

tetha
fraction of reservoir encircled by the aquifer, degrees. Used in 'uss_rad_edge', 'pss_rad_edge', and 'pot' aquifer models

muw_a
aquifer water viscosity, cp

cw_a
aquifer water compressibility, a numeric value, 1/psi

cf_a
aquifer formation compressibility, a numeric value, 1/psi

wf
weight factor, a numeric vector of zeros and ones. A zero value excludes the entire row of reservoir history data at a particular time from the material balance analysis

sgrw
residual gas saturation in water invaded zone (aquifer encroachment or water injection), a numeric fraction

Value

a list of class ’mbal_gas’ with all the required parameters for the mbal_perform_gas() S3 methods